Refactor: Modularized core services, engineered manual Macro engine & expanded node arsenal
Dismantled the monolithic voice-api structure to enforce clean separation of concerns. System functions are now cleanly segregated into domain-specific modules (files, gmail, keyboard, gallery, notes, coding), creating a highly scalable and maintainable architecture for future tool integration.

Core System & UI Upgrades:
- Manual Ignition Engine: Engineered a native, deterministic execution loop within the Workflow Editor. Introduced a manual 'RUN' button to bypass AI interpretation, allowing for instant, latency-free physical testing of macro sequences via direct command routing.
- Pure Execution Logic: Synchronized the visual builder and the voice assistant to use a mirrored, robust if/else execution pipeline, ensuring 100% predictable system hardware manipulation.
- Arsenal Expansion: Injected heavy-duty capabilities into the node schema, including full Email lifecycle management (read, draft, send), local Wormhole tunneling (deploy, close), and advanced temporal WhatsApp scheduling.
- UI Polish: Mapped custom vector icons and dynamic color states for the newly added Communication and Server nodes to maintain the cyberpunk aesthetic.
